requirements.txt
# Core dependencies
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
confuse==1.5.0
decorator==5.1.1
exceptiongroup==1.2.2
ffmpeg==1.4
idna==3.7
iniconfig==2.0.0
joblib==1.4.2
loguru==0.7.2
mediafile==0.8.0
msgpack==1.0.8
mutagen==1.47.0
mysql-connector-python==9.0.0
numpy==2.0.1
packaging==24.1
pandas==2.2.3
platformdirs==4.2.2
PlexAPI==4.15.15
pluggy==1.5.0
pycparser==2.22
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
threadpoolctl==3.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.2

# Local BPM analysis (librosa) - requires Python <3.13
# Will be replaced with essentia in Phase 2
# audioread==3.0.1
# lazy_loader==0.4
# librosa==0.10.2.post1
# llvmlite==0.43.0
# numba==0.60.0
# pooch==1.8.2
# scikit-learn==1.5.1
# scipy==1.13.1
# soundfile==0.12.1
# soxr==0.4.0
